make code more dynamic

+++ frontend/src/app/service/token-interceptor.service.ts
@@ -6,15 +6,18 @@ import {
   HttpEvent,
   HttpErrorResponse,
 } from '@angular/common/http';
-import { Observable, throwError } from 'rxjs';
-import { catchError, switchMap } from 'rxjs/operators';
+import { BehaviorSubject, Observable, throwError } from 'rxjs';
+import { catchError, filter, switchMap, take } from 'rxjs/operators';
 import { TokenService } from './tokenservice';
 import { Router } from '@angular/router';
 
 @Injectable()
-export class TokenInterceptor implements HttpInterceptor {
-  constructor(private tokenService: TokenService, private router: Router) { }
 
+export class TokenInterceptor implements HttpInterceptor {
+  private refreshingToken: boolean = false; // Flag to check if a refresh is in progress
+  private refreshTokenSubject: BehaviorSubject<any> = new BehaviorSubject<any>(null);
+
+  constructor(private tokenService: TokenService, private router: Router) { }
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     const authData = this.tokenService.getAuthData();
@@ -26,11 +29,8 @@ export class TokenInterceptor implements HttpInterceptor {
     }
 
     return next.handle(request).pipe(
-      catchError((error) => {
-        // console.error('Request error:', error);
-
-        if (error instanceof HttpErrorResponse && error.status === 401 && refreshToken) {
-          console.log('Handling 401 Unauthorized error with refresh token.');
+      catchError((error: HttpErrorResponse) => {
+        if (error.status === 401 && refreshToken) {
           return this.handleUnauthorizedError(request, next, refreshToken);
         }
         return throwError(error);
@@ -39,21 +39,32 @@ export class TokenInterceptor implements HttpInterceptor {
   }
 
   private handleUnauthorizedError(request: HttpRequest<any>, next: HttpHandler, refreshToken: string): Observable<HttpEvent<any>> {
-    return this.tokenService.refreshAccessToken(refreshToken).pipe(
-      switchMap((refreshResponse) => {
-        this.tokenService.setAccessTokenInCookie(refreshResponse.accessToken, refreshToken, JSON.stringify(this.tokenService.getAuthData()?.userInfo || ''));
-        const newRequest = this.addAuthorizationHeader(request.clone(), refreshResponse.accessToken);
-        return next.handle(newRequest);
-      }),
-      catchError((refreshError) => {
-        //        console.error('Error during token refresh:', refreshError);
-        this.tokenService.removeAuthData();
-        // this.router.navigate(['/login']);
-        return throwError(refreshError);
-      })
-    );
-  }
+    if (this.refreshingToken) {
+      return this.refreshTokenSubject.pipe(
+        filter(token => token != null),
+        take(1),
+        switchMap(() => next.handle(this.addAuthorizationHeader(request, this.tokenService.getAuthData()?.accessToken || '')))
+      );
+    } else {
+      this.refreshingToken = true;
+      this.refreshTokenSubject.next(null);
 
+      return this.tokenService.refreshAccessToken(refreshToken).pipe(
+        switchMap((refreshResponse) => {
+          this.refreshingToken = false;
+          this.refreshTokenSubject.next(refreshResponse.accessToken);
+          this.tokenService.setAccessTokenInCookie(refreshResponse.accessToken, refreshToken, JSON.stringify(this.tokenService.getAuthData()?.userInfo || ''));
+          return next.handle(this.addAuthorizationHeader(request, refreshResponse.accessToken));
+        }),
+        catchError((refreshError) => {
+          this.refreshingToken = false;
+          this.tokenService.removeAuthData();
+          this.router.navigate(['/login']);
+          return throwError(refreshError);
+        })
+      );
+    }
+  }
 
   private addAuthorizationHeader(request: HttpRequest<any>, token: string): HttpRequest<any> {
     return request.clone({
@@ -62,4 +73,4 @@ export class TokenInterceptor implements HttpInterceptor {
       },
     });
   }
-}  
+}
